refactored the parser into seperate modules
